requirements.txt
fastapi
uvicorn
streamlit
transformers
torch
faiss-cpu
requests
